GH-44 - Fallback to the naive onchain voting if anything goes wrong with offchain voting (#68)
* wip fallback

* remove moloch v2, merge master

* change the way we configure things

* fix test

* reorg code

* fix tests

* fix tests

* no more adapter dependency within core

* add check when you publish a vote result

* fix lint

* fix path

* fix path

* use node js 10 instead of 12

* reverting node version

* try to fix test on CI

* remove specific solc version

* use specific solc

* get weights from the dao directly to avoid weight issues

* get block number right before the vote

* const -> let

* blockNumber mismatch was the issue

* remove console

+++ contracts/test/OLToken.sol
@@ -2,10 +2,12 @@ pragma solidity ^0.7.0;
 
 // SPDX-License-Identifier: MIT
 
-import "../v3/utils/ERC20.sol";
+import "../utils/ERC20.sol";
 
 contract OLT is ERC20 {
-    constructor (uint _totalSupply) ERC20("OpenLawToken", "OLT", _totalSupply) {
-        _mint(msg.sender, 10000 * (10 ** uint256(decimals())));
+    constructor(uint256 _totalSupply)
+        ERC20("OpenLawToken", "OLT", _totalSupply)
+    {
+        _mint(msg.sender, 10000 * (10**uint256(decimals())));
     }
 }
